requirements.txt
annotated-types==0.7.0
anyio==3.7.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
fastapi==0.104.1
h11==0.16.0
idna==3.11
PyAudio==0.2.14
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
pydub==0.25.1
python-dotenv==1.2.1
requests==2.32.5
sniffio==1.3.1
srt==3.5.3
starlette==0.27.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.24.0
vosk==0.3.45
websockets==15.0.1
